refactor: tour service
