+++ src/app/app.module.ts
@@ -9,7 +9,6 @@ import { getAuth, provideAuth } from "@angular/fire/auth";
 import { IonicModule, IonicRouteStrategy } from "@ionic/angular";
 
 // Libs
-import { ColorSketchModule } from "ngx-color/sketch";
 import { LottieModule } from "ngx-lottie";
 import player from "lottie-web";
 import { NgxMatomoTrackerModule } from "@ngx-matomo/tracker";
@@ -24,11 +23,10 @@ import { AppComponent } from "./app.component";
 import { AppRoutingModule } from "./app-routing.module";
 import { SharedModule } from "./shared/shared.module";
 import { environment } from "src/environments/environment";
-import { ThemeEditorComponent } from "src/app/feature/theme/theme-editor/theme-editor.component";
-import { TourComponent } from "./feature/tour/tour.component";
 import { httpInterceptorProviders } from "./shared/services/server/interceptors";
 import { TemplateComponentsModule } from "./shared/components/template/template.module";
 import { ContextMenuModule } from "./shared/modules/context-menu/context-menu.module";
+import { TourModule } from "./feature/tour/tour.module";
 import { ErrorHandlerService } from "./shared/services/error-handler/error-handler.service";
 
 // Note we need a separate function as it's required
@@ -38,7 +36,7 @@ export function lottiePlayerFactory() {
 }
 
 @NgModule({
-  declarations: [AppComponent, ThemeEditorComponent, TourComponent],
+  declarations: [AppComponent],
 
   entryComponents: [],
   imports: [
@@ -52,11 +50,11 @@ export function lottiePlayerFactory() {
     provideFirebaseApp(() => initializeApp(environment.firebaseConfig)),
     provideAuth(() => getAuth()),
     FormsModule,
-    ColorSketchModule,
     LottieModule.forRoot({ player: lottiePlayerFactory }),
     // NOTE CC 2021-11-04 not sure if cache causes issues or not https://github.com/ngx-lottie/ngx-lottie/issues/115
     // LottieCacheModule.forRoot(),
     TemplateComponentsModule,
+    TourModule,
     NgxMatomoTrackerModule.forRoot({
       siteId: environment.analytics.siteId,
       trackerUrl: environment.analytics.endpoint,
