+++ src/app/feature/tour/tour.module.ts
@@ -1,0 +1,20 @@
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { FormsModule } from "@angular/forms";
+import { RouterModule } from "@angular/router";
+import { IonicModule } from "@ionic/angular";
+
+import { TourComponent } from "./tour.component";
+
+@NgModule({
+  imports: [
+    IonicModule,
+    CommonModule,
+    FormsModule,
+    RouterModule.forChild([{ path: "", component: TourComponent }]),
+  ],
+  exports: [],
+  declarations: [TourComponent],
+  providers: [],
+})
+export class TourModule {}
